# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MoodBite API",
    "description": "Gợi ý nhà hàng theo ngữ cảnh: câu tìm kiếm tự do + vị trí + thời điểm.\n\nMọi response bọc trong `data` (thành công) hoặc `error` (lỗi).\nXem GET /api/v1/health để biết nguồn dữ liệu nào đã sẵn sàng.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/me/collections": {
      "get": {
        "tags": [
          "me"
        ],
        "summary": "List Collections",
        "description": "Mọi bộ sưu tập của chính chủ, KÈM mục bên trong và số mục. Bộ mới nhất đứng đầu.\n\nTrả luôn mục trong một lượt: số bộ bị chặn ở 50 và số mục mỗi bộ ở 200, nên gọi thêm\nmột request cho từng bộ chỉ tốn lượt mạng mà không tiết kiệm được gì đáng kể.",
        "operationId": "list_collections_api_v1_me_collections_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionsResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Create Collection",
        "operationId": "create_collection_api_v1_me_collections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/collections/{collection_id}": {
      "patch": {
        "tags": [
          "me"
        ],
        "summary": "Rename Collection",
        "operationId": "rename_collection_api_v1_me_collections__collection_id__patch",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectionNameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Delete Collection",
        "description": "Xoá bộ và MỌI mục trong bộ. Các mục vẫn còn nguyên ở \"Yêu thích\"/\"Đã lưu\" nếu có —\nbộ sưu tập chỉ là một cách nhóm, không sở hữu dữ liệu đã lưu.",
        "operationId": "delete_collection_api_v1_me_collections__collection_id__delete",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/collections/{collection_id}/items": {
      "post": {
        "tags": [
          "me"
        ],
        "summary": "Add Collection Item",
        "description": "Thêm một quán/món vào bộ. IDEMPOTENT: thêm lại thứ đã có chỉ cập nhật tên.\n\nTrả 200 (không phải 201) kèm CẢ BỘ sau khi thêm: thao tác này không tạo ra tài nguyên\ncó địa chỉ riêng, và client cần số mục mới để cập nhật giao diện.",
        "operationId": "add_collection_item_api_v1_me_collections__collection_id__items_post",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCollectionItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/collections/{collection_id}/items/{item_type}/{item_id}": {
      "delete": {
        "tags": [
          "me"
        ],
        "summary": "Remove Collection Item",
        "operationId": "remove_collection_item_api_v1_me_collections__collection_id__items__item_type___item_id__delete",
        "parameters": [
          {
            "name": "collection_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Collection Id"
            }
          },
          {
            "name": "item_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Type"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "COLLECTION_NOT_FOUND | ADDRESS_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "DATA_NOT_READY",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHORIZED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddCollectionItemRequest": {
        "properties": {
          "item_type": {
            "type": "string",
            "title": "Item Type",
            "description": "restaurant | dish"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Tên để hiển thị, chụp lại lúc thêm."
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "item_id",
          "name"
        ],
        "title": "AddCollectionItemRequest"
      },
...
      "CollectionItemSchema": {
        "properties": {
          "item_type": {
            "type": "string",
            "title": "Item Type",
            "description": "restaurant | dish"
          },
          "item_id": {
            "type": "string",
            "title": "Item Id"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Tên chụp lại lúc thêm vào bộ."
          },
          "added_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added At"
          }
        },
        "type": "object",
        "required": [
          "item_type",
          "item_id",
          "name"
        ],
        "title": "CollectionItemSchema"
      },
      "CollectionNameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Tên bộ sưu tập, 1–60 ký tự."
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CollectionNameRequest"
      },
      "CollectionResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionSchema"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CollectionResponse"
      },
      "CollectionSchema": {
        "properties": {
          "collection_id": {
            "type": "string",
            "title": "Collection Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "item_count": {
            "type": "integer",
            "title": "Item Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/CollectionItemSchema"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "collection_id",
          "name",
          "item_count",
          "items"
        ],
        "title": "CollectionSchema"
      },
      "CollectionsData": {
        "properties": {
          "collections": {
            "items": {
              "$ref": "#/components/schemas/CollectionSchema"
            },
            "type": "array",
            "title": "Collections"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "collections",
          "total"
        ],
        "title": "CollectionsData"
      },
      "CollectionsResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CollectionsData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "CollectionsResponse"
      },
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageData": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Câu thông báo hiển thị thẳng cho người dùng."
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "MessageData",
        "description": "Kết quả cho các thao tác không trả về tài nguyên nào."
      },
      "MessageResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageData"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "MessageResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}